Begin tidy up of TGenericBigInteger operators, lots of unnecessary inlines
There were far to many overly-complicated inline functions.  This patch
begins the reimplementation of, for example add(), operator+=() and
operator+() so that

 - add() is removed
 - operator+=() does the work in place
 - operator+(x) is { return TGenericBigInteger<tLittleInteger>(*this) += x; }

There was no need for add(), and it simply obscured matters since it was
a simple wrapper around operator+=(); with *this as one of add()'s
arguments.  That required that add() new how to cope when one of the
inputs was 'this' -- now it is guaranteed that one of the inputs is
'this', which makes that test unnecessary and we can code for the single
case.  The time when a temporary is necessary is now pushed to the
correct place: the caller rather than the callee.

The same idea is begun for many of the other operators, but the
refactoring of the functions isn't done yet.