Query a virtual function to decide on command code acceptance
Rather than have TMessage::queryMessageExtractSize() using a hard coded
decision method (string comparison) to decide whether the offered string
contains a message of a type that matches the current class; this patch
moves that decision to a new virtual member function.

That virtual, by default, does the same string comparison that TMessage
always did before, but it offers the possibility to not use that method.

TMessageUnimplemented now takes advantage of that facility to override
acceptCommandCode(), making it accept anything that looks like a
message.