Give the TMessage_version hierarchy a way of querying minimum accepted version
There have been multiple versions of the version message.  We want to be
able to read all of them.

TMessage_version contains the common code for all versions of the
message (or will).  Version 0 of the message is the one that reads the
version number from the message and checks it for suitability, so
TMessage_version_0 supplies a virtual that the child classes override to
let it know whether they consider it acceptable.