Implement parse() methods for the TMessage_version hierarchy
The hierarchy is simple; with each child class calling its base class to
parse the backward compatible parts.

 - TMessage_version_209::parse() calls TMessage_version_106::parse()
 - TMessage_version_106::parse() calls TMessage_version_0::parse()
 - TMessage_version_0::parse() calls TMessage_version::parse()

TMessage_version represents _all_ versions, and so it simply clears all
the fields it knows about ready for the child to fill in.  It also
supplies a printOn() to allow human-readable display of the message.