Add a UNITTEST for running sample messages through the template list
Each sample message is offered to each template message.  The first that
accepts it is clone()d and then the message passed to it for full
parsing.  That parse() can still fail if the message has faults in it,
but those faults are message-type-specific.  If one of those occurs the
error is published and the clone deleted.