Implement TMessage::className() and clone()
className() is simply for user information, it's handy to have a base
class pointer be able to tell the user (or in the absence of
dynamic_cast<> the program) what the object really is.

clone() is simply the dynamic memory version of a copy constructor.  It
allows us to copy an object using a base class pointer.  i.e. we don't
need to know what the derived type is:

  TBase *BasePointer1 = new TDerivedType;
  TBase *BasePointer2 = BasePointer1->clone();

The object pointed to by BasePointer1 is copied to a new object, and
BasePointer2 points at it; the copy is of a TDerivedType object, but the
copy is performed using a TBase pointer.