Add message digest calculation support to crypto module
The TSSLMessageDigest class acts as a base class to represent any hash
available in OpenSSL.  Child classes are instantiated from a template
with the appropriate SSL structure returned from the appropriate class
virtual.

TSSLMessageDigest itself wraps up the (slightly complex) OpenSSL calls
needed to use the hashes.  Making hashing as simple as:

 THash_sha256 SHA256;
 string digest = SHA256.transform( message );

Which is considerably simpler than calling OpenSSL directly.

A UNITTEST is included with samples from the bitcoin protocol
specification page (not all of which are correct, I believe).