Use new TMessageDigest base class to supply TDoubleHash
TDoubleHash accepts two other TMessageDigest objects during
construction.  Those objects are then called in its transform()
implementation in order.

   DoubleHash = Hash2(Hash1(plaintext))

This makes it trivial to change from a single hash to a double hash, of
whatever mixture is wanted.

That facility is used to make SHA256(SHA256()) and RIPEMD160(SHA256())
in the UNITTEST.

Finally, the checksum implementation in TMessageWithChecksum is updated
to use TDoubleHash rather than needing to explicitly double call SHA256.