Create a common base class for string-based message elements and add child classes of that type
The three string types are:

 - TNULTerminatedStringElement; reads from the stream until reaching a
   NUL character.
 - TSizedElement; reads N characters from the stream, with N given to
   the constructor.
 - TFixedSizeElement; reads N characters from the stream with N given to
   a template parameter.  This template creates a simple wrapper class
   around TSizedElement.  The advantage is that a template parameter can
   be specified with the declaration rather than needing to create
   another constructor in whatever structure is holding the
   TSizedElement.

e.g.

  struct sStructure {
    sStructure() : StringField(10) {}
    TSizedElement StringField;
  }

Can be replaced with

  struct sStructure {
    TFixedSizeElement<10> StringField;
  }

I believe the code overhead will be zero if the compiler's optimiser is
doing its job.  It should notice that TFixedSizeElement<>'s constructor
is simply a call to TSizedElement's constructor with a literal constant,
and optimise out the intermediate call.  Even if it doesn't -- I'll live
with a slightly inefficient initialisation that happens rarely.