Refactor TMessage hierarchy to use istream with new TMessageElements instead of strings
This is quite an upheaval, but the idea is to simplify the job of
implementing all the (as yet unimplemented) message types.  Instead of
putting message specific code in each and every parse(), the member
elements of each message are represented by TMessageElements and read
really easily like this:

  inputstream >> MessageElement1 >> MessageElement2
  if( SomeTestOn(MessageElement2) )
    inputstream >> MessageElement3

The logic of the message is captured, but the detail of how each field
is read can be deferred to the appropriate TMessageElement, which knows
how to read itself.