Remove TMessage::queryMessageExtractSize(), and use istream + exceptions
This was still left over from the times when a string was the input to
TMessage; now it's istream and we can't rely on having a string
available.  Therefore we junk this member and rely on the parser itself
to throw exceptions as necessary.

This patch adds to the UNITTEST to show how to handle those exceptions.

Note in particular the use of tellg() and seekg() to allow the same
data to be fed to multiple templates.