Have lib/Makefile automatically produce the list of library dependencies
There is a .mak file for every module that needs a library.  These
contain specially named variables that allow the master makefile to
build and link the UNITTESTs.  Those same variables can be stored to a
library-wide .mak file for inclusion in the applications.

For example: crypto.mak might contain

   crypto_LIBS := crypto ssl

To include the ssl libraries when building with it.  This new makefile
recipe will create libbitcoin.mak which contains

   libbitcoin_LIBS := crypto ssl

Which in turn can be included in an application Makefile.