Add constants module to hold magic numbers, like seed node addresses
Magic numbers are those that have no particular meaning.  They are
simply needed to fill certain fields in or write in headers.

This patch fills the module with the seed node list from the official
client.  The official client uses an array of unsigned ints, I've
passed that list to an object array, with the object capable of
construction from an unsigned int.  This uses no more storage, but gives
us a place to put convenience functions that interpret the constants for
us.

A UNITTEST demonstrates how it works by printing all the address to the
terminal in dotted-quad form.