Make peer module and bitcoin network module, making TNodeInfo base to TOfficialSeedNode
TNodeInfo is now the base class for TOfficialSeedNode in the constants
module.  The peer module also hold TBitcoinPeer to represent a connected
peer.  The bitcoinnetwork module holds an array of these, and a
TNodeInfo array representing the accumulated information about any node
on the network, regardless of whether its connected or not.

It's expected that that directory will initially be populated from the
official seed node list, then as live connections become available, from
the addr messages we get from other nodes.

So: the naming convention I've chosen is:

 - Node = any host running the bitcoin client
 - Peer = a node that we ourselves are connected to.  It has a
   particular version to itself; and so gets a protocol factory to read
   the appropriate set of messages.
 - Network = a collection of nodes and peers (and eventually a
   blockchain)

In other words: a peer is created from a node (my intent is that a node
will be turned into a peer by a network).