Implement a per-version instance of TMessageFactory
The child classes of TMessageFactory each have an init() override that
fills the template array with the appropriate TMessage children for that
version.

TMessageFactory::receive() then iterates through only those templates.
That means, if the appropriate TMessageFactory is constructed for a
particular version of peer, then it will interpret the messages received
from it appropriately as well.