Add a 'template' flag to TMessage
Some TMessages are created only to act as templates to be added to a
collection that defines a version.  Unused at the moment, this patch
adds a flag to allow these template versions to be distinguished.

The template arrays are always kept as const pointers; so the flag is
marked mutable so that it can still be changed in those arrays.