Make TMessage_version child classes capable of creating the matching TMessageFactory
TMessageFactory descendent essentially store the list of TMessages that
a particular version of the protocol supports.

After a TMessage_version is received, we'll want to make the
TMessageFactory that matches the version in the packet.
TMessage_version::createMessageFactory() can be used to make the
appropriate TMessageFactory without needing to look in the packet
itself.