First pass at design of scripting classes
The three key classes are TExecutionStack, TBitcoinProgram and
TStackOperator.

Essentially, a TBitcoinProgram is an object that maintains a list of
TStackOperator which operate on a TExecutionStack.

TBitcoinProgram assumes a template list of TStackOperators is created by
a TBitcoinProgram child class, and tries each template object in turn
against an incoming opcode, gradually constructing the program list.

With this patch only a few of the operators are implemented for
execution, but all of them should be capable of reading themselves from
a stream.

The ones that do execute are those special cases: for example pushing
raw data onto the stack NOPs and invalid opcodes.