Move all the TStackOperators from the header to the module
There is no reason that any other module needs to know the
implementation details of the stack operators, other modules need to be
able to run the script, not change it.

This patch leaves TBitcoinProgram and TExecutionStack in the header, so
that other modules can load a program and get its result.