TMessageFactory::createVersionedBitcoinScript() returns a new TBitcoinScript instance of the appropriate version
TMessageFactory is the object that knows the right set of TMessages for
a particular protocol version, it's a natural extension to give it the
ability to process the scripts associated with that version by creating
the appropriate TBitcoinScript child.

The protocol spec doesn't actually list which set of operations went
with any particular version, so I'm assuming it's never changed.
Therefore every createVersionedBitcoinScript() returns TBitcoinScript_0;
but that's hardly the point.