Implement facility to make a single OP code represent multiple TStackOperators
A few Opcodes (OP_EQUALVERIFY, OP_NUMEQUALVERIFY, OP_CHECKSIGVERIFY,
OP_CHECKMULTISIGVERIFY) are simply compound representations of multiple
other Opcodes.  For example, OP_EQUALVERIFY is simply an OP_EQUAL
followed by an OP_VERIFY.

Presumably these compound representations were meant to allow scripts to
be shorter by a byte or two.  Whatever the reason, just because they are
shorter scripts doesn't mean we should duplicate code.  That is:
implementing OP_EQUALVERIFY should be done as a call to the already
implemented OP_EQUAL and OP_VERIFY, rather than copy-and-paste those
functions together.

This patch represents the solution to that problem.  The method of
reading opcodes is changed so that the read is now done by the template
TStackOperatorFromStream itself.  The template is given a new virtual,
readAndAppend(), which simply does what the old TBitcoinScript::read()
used to do: clone() and call the clone's read().  This leaves existing
functionality exactly as it was, although the work is moved internal to
the classes.

However, being a virtual means that some classes can override readAndAppend(),
those that do are the compound operators.  The override simply appends
whichever TStackOperators are appropriate to the given script.

This might look like unnecessary overkill for such simple operators; but
I have it in mind to use the same technique for implementing
OP_CHECKSIG, which I believe can be better expressed as a compound
operator.