Implement lots of TStackOperators
This patch implements all the easy operators; basically anything that
can be done in a few lines of C++ is now done.  This includes:

 - Stack manipulators
 - Arithmetic
 - Boolean Logic
 - Bitwise logic
 - String manipulation

Significantly, none of the cryptographic operators have been
implemented, and further: the logic and arithmetic are all done on
standard integers.  It's almost certain that they should be done using
arbitrary sized integers.

To implement that requires linking in with the extraint module after
I've added support for signed values.  The code here won't change (other
than maybe a change of TStackElementInteger's underlying Data type), so
this will all do fine as it is.