Explicit template instantiations must come after the template function definitions
It's not unreasonable, but I wasn't aware of it.  It seems that G++
requires that explicit template instantiations must come after the
template function definitions; not just after the template declaration.

For example, consider this header:

   // Template declaration
   template <typename T> class Object {
      void somefunction(T);
   }

   // Forward explicit template declaration instantiation
   extern template class Object<int>;

If we perform the explicit template instantiation in the module file
before the definition:

   template class Object<int>;

   template <typename T>
   void Object<T>::somefunction(T) {}

Then Object<int>::somefunction(int) is not instantiated and is never
exported as a symbol.

This was showing up as the script module giving linking errors against
the TBigInteger unary operators.  It was nothing to do with the
operators though.  It was because the explicit instantiation should be
done like this:

   template <typename T>
   void Object<T>::somefunction(T) {}

   template class Object<int>;

Which is what this patch changes in both extraint module and the crypto
module (although it didn't make any difference in the crypto module).