Give TBlockHeaderElement::DifficultyBits its own sub element, TDifficultyTargetElement
The docs say that this is a thirty two bit number; but when you look
deeper, the page on difficulty bits says that the difficulty stored in
the header of the form

  0xaabbccdd  (stored in the stream as DD CC BB AA)

Is actually to be used as

  DifficultyTarget = 0xbbccdd * 2^(8*(0xaa - 3))

So it's certainly not a simple 32 bit unsigned int.  It is in fact
closest to being a 24 bit little endian mantissa and an 8-bit offset
exponent.

This target represents a 256 byte number, since multiplying by 2 to the
power something is a shift left, the highest this can shift is 2016
bits, with the three bytes of mantissa defining the top 24 bits.