The operator~() implementations in extraint were returning the wrong object
The operator~() returns a new object (or should), representing the
inversion of the bits in the called-on object.

The previous implementation was returning the called-on object and
effectively discarding the calculated object.

That was actually fortunate because the target digits array wasn't being
cleared before use which would insert an extra zero block at the least
significant end.