Give coins their own class to represent them
I suspect that it's very quickly going to become annoying to deal with
the convoluted maths of representing "HundredsOfNanoCoins", so I've
given them their own class to encapsulate that maths as it becomes
necessary.

As an example: the setting and getting of the coins value in "coins"
rather than "nanocoins" (which aren't nanocoins anyway) is now hidden in
two helper functions.