TGenericBigInteger::fromBytes() was clipping the most significant byte
The safety check in fromBytes() checked that the calculated string index
was greater than zero, but of course zero is a valid string index, so
the first byte of any string of bytes was being discarded.

The fix is to change the check to greater than or equal to.

Oops.