TBitcoinScript now keeps two template lists: authorisation and claimant operators
https://en.bitcoin.it/wiki/Protocol_rules says

"7. Reject 'nonstandard' transactions: scriptSig doing anything other
than pushing numbers on the stack, or scriptPubkey not matching the two
usual forms"

This patch makes that possible by having two template lists in a
TBitcoinScript, one for claimants (scriptsig) and one for authorisation
(scriptPubkey).  TBitcoinScript_0::init() is updated to put the
appropriate operators in each list in accordance with the above rule and
TBitcoinScript::read() gets an extra parameter to specify which list
should be used.