Start storing bitcoin 'constants' in TNetworkParameters
The official bitcoin client has a number of rules that it enforces.
Many of those rules are numeric specifications about particular message
attributes.  Rather than have them as literal constants, I've chosen to
treat them as network parameters.  This is because I can well imagine a
situation where it would be convenient to introduce a rule change in the
testnet without breaking the production network rules.

Other than the fact that they aren't constants, they will be treated for
all other purposes as if they were.