Add constants for specifying a COINBASE transaction and use them
A COINBASE transaction has no source transaction; and is indicated by
the special case of setting the output hash reference to zero and the
output reference index to -1.

However, these special numbers aren't self-explanatory when used
literally, so this patch adds constants for them and uses them in
TPredefinedNetworkParameters::configureGenesisMessage().