TMessage_block and TMessage_tx have versions, supply a virtual to query it
`block` and `tx` messages have a version field.  At present their is
only one version of these messages, but should there be more in the
future we need a way of handling it.  That way will be TMessage_block
and TMessage_tx child classes, which will override the new virtual
supplied so that externals can find out what sort of class to create to
represent a particular block.

For example, if, in the future there is a version 2 block; we might then
have a database with some version 1 blocks and some version 2 blocks; we
will want a way of creating an appropriate TMessage_block from these
database records.  We'll do that by calling the
minimumSupportedVersion() member in an array of templates, just as we do
for the TMessage_version children.

As I say, it's a moot point at present, there are only version 1 blocks
and transactions.