Add TNetworkParameters::limitDifficultyTimespan()
Wherever possible I'm going to try to keep arbitrary numerical limits
and checks as part of the network parameters.  There is no particular
reason why limiting difficulty updates to a factor of four is any better
than limiting to a factor of two.  That sort of choice should therefore
be centralised.

To allow other values to be tested (should anyone want to) without
affecting the production network, that centralisation should obviously
be in the TNetworkParameters class.