Add TBlock::getNextRequiredDifficulty()
The difficulty required of a proof of work for a block is normally the
difficulty of the previous block.  However, every N blocks (with N set
by the network parameters), that required difficulty is reevaluated.

It is reevaluated since the number of generating nodes in the network is
not known ahead of time.  That means the total hashing power is unknown.
If more nodes are added, then the blocks will start to be generated
faster.  Therefore the network must decide to make it harder to generate
blocks in order to keep the rate of block generation constant.

Every N blocks then, the time taken to generate those N blocks is
compared with the target time (again a network parameter) for generating
blocks (it's 2016 blocks every two weeks at time of writing).  The
difficulty is then adjusted by the factor that the observed time
deviates from the target time.

For example: if it took 3 weeks to generate 2016 blocks, then the
difficulty can be assumed to be 2/3 of what it should be, and so should
be decreased by 2/3.  If it took 1 week to generate 2016 blocks, then
the difficulty is assumed to be half what it should be and should be
increased by 2/1.

 New_Difficulty *= TARGET_TIME / OBSERVED_TIME

In the case of this patch, the difficulty is represented by the target
threshold rather than the difficulty, and for our purposes, difficulty
and target are reciprocals, so:

 New_Target *= OBSERVED_TIME / TARGET_TIME

Which, along with some safety checks is all
TBlock::getNextRequiredDifficulty() does.