Add TNetworkParameters::convertTargetToDificulty()
Difficulty is a floating point representation of the current target.  It
is calculated as:

  Difficulty = MaximumTarget / CurrentTarget

With MaximumTarget being the easiest because it is the biggest.

This patch cheats (as does the official client) and uses only the top 32
bits of the targets for calculating the value.  In fact, we do slightly
better than the official client which does this:

    int nShift = 256 - 32 - 31; // to fit in a uint
    double dMinimum = (CBigNum().SetCompact(bnProofOfWorkLimit.GetCompact()) >> nShift).getuint();
    double dCurrently = (CBigNum().SetCompact(pindexBest->nBits) >> nShift).getuint();
    return dMinimum / dCurrently;

The thing here is that it doesn't just use the "best" 32 bits, it forces
the use of the top 31 bits, after having discarded all but the top three
bytes (by converting to-and-fro between the compact representation)

For example, if we start with:

    255 ------ 193 ----------- 0
      XXXXXXXXXXXXxxxxxxxxxxxxxx

Then after shifting we end up with

     63 -------- 0
      XXXXXXXXXXXX

However, by using special knowledge that the MaximumTarget is 32 bits of
zero followed by all ones, and the guarantee that CurrentTarget is less
than the MaximumTarget, we can be sure that it actually looks like this:

     63 ------- 32 ----------- 0
      000000000000XXXXXXXXXXXXXX

i.e. 31 bits.

This is all very well, until we remember that the proof of work limit is
different for the test network.  Therefore the official client will get
the difficulty wrong for the test network.

This patch doesn't hard code any bit positions.  Instead it uses the
"highestBit()" function of TGenericBigInteger and simply preserves the
as many bits as is possible before performing the division.  This should
give a more accurate estimate.