Add Checkpoints map to TNetworkParameters
The chain already exists "out there", in an agreed upon form.  That
means there are blocks with known hashes and heights out there.  This
patch adds the ability to list those height and hash joins in
TNetworkParameters, so that the block chain can check itself as it's
being built.

I've added the checkpoint list from the official client to
TProdnetNetworkParameters.