Convert KNOWN_NETWORKS to be a singleton
We want the known networks to be initialised on demand: all of them at
once.  Therefore the known networks must be represented by a single
class that we can declare a singleton.

This patch does just that.  The KNOWN_NETWORKS class is really just a
wrapper around a set<> of TNetworkParameters pointers and a constructor
that creates TTestnetNetworkParameters and TProdnetNetworkParameters
instances.

TSingleton is then used to wrap KNOWN_NETWORKS and takes care of the
singleton creation on demand.

This is necessary so that the initialisation order is more controlled.
Using module globals makes initialisation order essentially random, at
the whim of the compiler and linker; on demand initialisation makes the
creation order automatically be whatever is required.