Fix packet synchronisation faults in TMessageFactory::receive()
The magic number defined by the network parameters can be used to find
the beginning of the next packet.  Rather than try every single byte as
a start byte, the factory can now use the magic bytes to sync.  This
also allows it to know where the previous packet ends and so discard it.

In short: the factory will now cope with invalid or incorrect packets
and will resync as quickly as it can.