Use string::c_str() for printing command bytes to prevent NULs getting to the terminal
The command string in the header block is a fixed size and is padded
with NULs.  string being string, it doesn't care about NUL bytes and so
happily passes them to the ostream used to print the command on screen.

There is no major problem, but printing non-printable characters,
particularly NUL to a terminal is pretty poor form.  This patch switches
to using string::c_str() so that the string is only printed up to the
first NUL by ostream.