Automatically setPeer() and setFields() in TBitcoinPeer::queueIncoming() and queueOutgoing()
Messages added to the peer's queues can be assumed to be owned by the
TBitcoinPeer and so can automatically have a setPeer() called.  By the
same token, once it's been pushed to the queue the creator has finished
with it, so we can assume they've done all they want to it and can call
setFields() as well.