Move all message processing to TBitcoinNetwork from TMessageFactory
Factories are no longer responsible for answering messages.  They were
never the appropriate place.  It is the network that will be able to
make decision about blocks and transactions, not factories, who's only
job is to convert byte streams to TMessages.

The handshaking is now done explicitly in TBitcoinPeer, which keeps
track of whether the remote has given a verack response, and whether the
local has sent a version message; only allowing the handshaking state to
be exited when this is true.

A particular change is that verack is now handled explicitly by a
_versioned_ message factory.  This is because we shouldn't make
assumptions about what form the verack response takes.  Once we know the
remote version, the appropriate factory is loaded, and hence the
appropriate verack handler.  While it doesn't matter at the moment, this
allows us to handle future protocols were verack has been changed.