Forgot to correct for non-full bytes in TGenericBigInteger::reversedBytes()
We're doing a byte reversal, not a block reversal in reversedBytes().  A
block reversal is a beginning, but if the highest block isn't full then
the least significant bytes will be wrong.

The solution is to apply a right bit shift at the end to compensate for
a non-full block.