TGenericBigInteger::operator>>= was broken for whole block shifts
The code in question always assumed that the adjacent blocks would need
modifying.  As they do for any bit shift other than one.

  underflowbits = B << (0)
  B >>= 0
  B |= lastbits
  lastbits = underflowbits

Here, underflowbits is equal to B, B is left alone (as it should be),
then B is ORed with lastbits, all fine.  Except that lastbits is then
set to underflowbits, which isn't zero but should be.  There are no
underflow bits when b is zero.  That meant that block zero would be ORed
with block one; block one would be ORed with block two, etc.