Fix Merkle tree calculation, the hashes are the reversed hash of the reversed hashes
Comparison with what blockexplorer showed was the Merkle tree told me
that I wasn't hashing the right thing.  It was fairly obvious that, as
everywhere else in bitcoin, the hashes are treated as being little
endian but displayed big endian.

So, when hashing two hashes, the two hashes need reversing before being
hashed, then the resulting hash is reversed for storage.