Use THashElement to take care of hash byte reversals in TMessage_block::calculateMerkleTree()
Previously all the hash reversals were done in place in
TMessage_block::calculateMerkleTree().  That was fine, but it obscured
the purpose of the Merkle tree calculator -- why should it repeat code
that already exists for taking a bitcoin hash and serialising it.  That
is the job of THashElement.

THashElement writes a bitcoin hash reversed, and reversed and hash it
reads.  Therefore when we use it we can remove all mention of reversals
from calculateMerkleTree(), just as it should be.  From the point of
view of someone reading the code, they need not be confused that the
reversal is part of the calculation.  It is necessary, but is just an
oddity of bitcoin's chosen method of hash serialisation.