TBlockPool::receiveInventory() now adds blanks to the pool
When an inventory item is received, we can assume that the object exists
(otherwise we wouldn't be offered it, and it's impossible to fake a
hash); therefore we can reasonably make space for it in the pool.

At the moment a NULL is added, but in the future we can add an empty
item and set a "pending" state on it.