TBitcoinPeer::receive() keeps the incoming bytes in a variable it can clear
When the factory changes from a TVersioningMessageFactory to a
TVersionedMessageFactory as state went from Handshaking to Connected,
the same bytes that were passed to the versioning factor were passed
to the versioned factory.  The problem is that the buffered bytes were
already copied internally from one factory to the other.

Now: we can't simply drop the factory's internal buffer because it can
contain more than just this iterations incoming bytes, particularly for
long messages.  Let's say that a block (which are typically tens of
kilobytes) is the first message received after the version message.
That means that thousands of bytes are needed to be sent to the factory
before a message can be made, which might be sent in 1000 byte blocks
over a number of calls to TBitcoinPeer::receive().  Well then, the
bytes received before the current call must be preserved as precious.

The answer then is that we do copy the internal buffer, but we don't
send this rounds incoming bytes multiple times.  That means we clear the
buffer when it's been used.