Add a TByteArray type which simply wraps vector<unsigned char>
TByteArray is a convenience class to add all the casts and operators
needed to make vector<unsigned char> easily usable where strings and C
arrays were previously used.

I'm changing because I read this:

http://herbsutter.com/2008/04/07/cringe-not-vectors-are-guaranteed-to-be-contiguous/

"Why is it so important that vectors be contiguous? Because that's what
you need to guarantee that a vector is layout-compatible with a C array,
and therefore we have no reason not to use vector as a superior and
type-safe alternative to arrays even when we need to exchange data with
C code. So vector is our gateway to other languages and most operating
systems features, whose lingua franca is the venerable C array."

Clearly then, the C++ designers intended vector<> to be the replacement
for character arrays _not_ string as I have been using.  I'll start
migrating to TByteArray where appropriate.