TEllipticCurveKey::sign() wasn't handling signatures shorter than the maximum
The storage for a signature has some maximum, and
TEllipticCurveKey::sign() was allocating that maximum.  However, the
fact that ECDSA_sign_ex() returns a signature length once it has seen
the real data implies that it's possible for the signature to end up
smaller than that maximum.  Therefore the buffer must be shrunk after
the signature has been generated to match what was returned.