bytearray module get TAutoClearAllocator for TByteArray
TAutoClearAllocator wipes the memory used to zero before freeing it.
This is intended for storage of secret keys, to make sure they are not
visible after they are finished with.

TByteArray gets custom allocator support so that it can take advantage
of this in the form of it becoming TByteArray_t.  The insecure and
secure versions are then typedefed to TByteArray and TSecureByteArray
respectively.  This leaves other code identical.