Add an (as yet uninitialised) nonce to TBitcoinNetwork
The only way we can really identify ourselves (given that we don't know
our external IP address) should we self-connect, is to have a unique
number we can check.

There is no way to create a guaranteed unique number, so we trust to the
laws of chance and choose a random 64-bit number.  That number is then
written in all outgoing version messages.  We then check what the
received nonce is, and if it matches our own we can be pretty sure that
we've connect to ourself.