Have TBitcoinPeer take a copy of the received version message
It's probable that the TBitcoinNetwork will want to be able to ask
questions about peer parameters.  We get that information from the
TMessage_version we receive early on.  Rather than create fields for
each of the version parameters, it's easier to just store a copy of the
version message we received.