Use TBitcoinNetwork to create the appropriate outgoing TMessage_version
While we try to be compatible with all versions, we do have our own
version.  TBitcoinPeer is version agnostic though, so we don't want to
create specific TMessage_version objects in it.

Instead we let TBitcoinNetwork supply whatever TMessage_version
derivative it wants (for now it simply returns TMessage_version_31402).