Move the TMessage_version response handler to TBitcoinNetwork from TBitcoinPeer
Once TBitcoinPeer knows what version the remote is, and the correct
message factory is loaded, it needs to have no more part in
communication.

However, it was being left to do a little bit of handling of the version
message.  This patch removes that and lets TBitcoinNetwork handle it.
TBitcoinPeer does this by pushing the version message it has received
back onto the incoming queue as soon as it has changed factories.  That
leaves the message to be pulled off the queue in the Connected state
handler and simply passed on to TBitcoinNetwork like any other packet.

TBitcoinNetwork::process() is then altered to return the appropriate
verack.