Change TMessageBasedBlock to use only the block header, and as a knock on move block hashing out of TMessage_block
The blocks are actually only a representation of the block headers; so
storing the whole message was overkill.  It also made it difficult to
fit the idea of headers-only mode into TBlock via receiveHeaders().

This patch is essentially one big cascade, caused by one small change:
TMessageBasedBlock::Message was a TMessage_block pointer; and is now a
TBlockHeaderElement called Header.  Everything else follows:

 - TBlock::updateFromMessage() is now updateFromHeader()
 - TBlock::getMessage() is now writeToHeader()
 - TBlock::getHash() can't call TMessage_block::calculateHash(), so must
   instead do it's hashing itself.  This actually isn't such a bad
   thing, it was always a toss up whether this should be in
   TMessage_block or TBlock; and having it in TBlock move TMessage_block
   closer to just being a message representation class rather than a
   block processing class (which is what TBlock is supposed to be).