Change TBlockPool::Branches from a set to a map of TBranches
The map key is the hash of the block representing the branch tip.  The
TBranch object represents per-branch data.  The per-branch data is made
up of:
 - A pointer to the currently being-mined block.  If we are generating,
   then we are trying to form the new branch tip.  That block is
   specific to that branch, so must be per-branch data.
 - A transaction queue.  Newly received transactions that have all their
   inputs available (and remember that "inputs available" is a question
   that can only be answered for a particular branch) are queued ready
   for adding to the mined block.  Since they might be already added to
   a block on another branch, they are a per-branch item.
 - Incomplete transaction pool.  Transactions that would go on the queue
   but can't because their inputs don't exist (yet).  It's possible that
   we receive a transaction before the transactions it depends upon.  We
   simply hold it until they turn up in the queue.
 - An adopted transactions pool.  These are the transactions that were
   on the queue, but have been moved to the mining block for inclusion.
   If we successfully mined a block then this is the list of transactions
   that go into that block.

The reason the transaction queue is per-branch may be seen if we
consider this chain:

  o -- o -- P -- 1 -- 2 -- branch1
             \
              3 -- 4 -- 5 -- branch2

Now imagine that we receive a new transaction, it isn't on either branch
already, therefore it is queued on both.  We have chosen branch1 as the
best branch, and we happen to generate a block.  That block then
includes that transaction.  Subsequently, branch2 generates a more
difficult block than us -- we want to switch our main branch to that
node.  That original transaction should still be in the branch2 queue,
which it is if we have stored the queue per-branch.