Don't destroy and create a new TVersionedFactory in TBitcoinPeer::receive() when we've already got one
It's possible to stay in the "Handshaking" state for more than one
iteration of messages (as we wait to receive a verack for example), even
after we have received a version message.  That meant that the version
message which is carefully pushed back onto the queue so that the
TBitcoinNetwork gets to see it once we enter the "Connected" state would
be handled multiple times -- with a new factory constructed, and the old
one (perfectly adequate one) deleted, as well as some log message noise.

This patch adds a check to only do this if we haven't already received
and processed a version message.