Oops: using TByteArray requires that we truncate the buffer after we know how many bytes were received
TByteArray is a C character array replacement; it's used for reading
like this:

   TByteArray ba(BUFFER_MAX);

   n = read( fd, ba, ba.size()
   ba.resize(n);

That final resize is necessary -- just like a C character array, writing
to the buffer itself doesn't tell it anything about how much data is in
it.  In C we would solve that problem by carrying around the buffer
length as well, in C++ we solve that problem by embedding the size into
the object that holds the buffer, TByteArray.

It isn't psychic though, if read() only returns 10 bytes, then
TByteArray must be informed that that is how long it is.

Without this, all the messages were appearing to be 1024 bytes long in
TBitcoinNetwork_Sockets::receiveFrom().